restaurantDetailContainer 테스트 통과
